% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{cdc_app_token}
\alias{cdc_app_token}
\alias{cdc_set_token}
\alias{cdc_get_token}
\title{Set and Get App token from CDC}
\usage{
cdc_app_token(token = NULL, quiet = FALSE)

cdc_set_token(token)

cdc_get_token()
}
\arguments{
\item{token}{App token from \href{https://data.cdc.gov/}{data.cdc.gov}. This should be a character string
which gets attached to your system environment for security.}

\item{quiet}{Boolean. Whether to include console outputs when running function.}
}
\value{
Character. The app token.
}
\description{
Helper function for setting and getting your CDC App token for use throughout the package.
This is not strictyly necessary as data.cdc.gov allows anonymous requests. App tokens are useful for
security reasons and to bypass the rate limits that are forced with anonymous API requests.
}
\details{
To receive your app token, you can go to \url{https://data.cdc.gov/signup}. After
setting up your account, you can go to the \href{https://data.cdc.gov/profile/edit/developer_settings}{developer settings} to
create a new app token.
}
\examples{
cdc_app_token()

cdc_app_token(token = "APP_TOKEN")

cdc_set_token(token = "APP_TOKEN")

cdc_get_token()

}
