% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cdc_browse}
\alias{cdc_browse}
\title{Open Dataset in Browser}
\usage{
cdc_browse(dataset_id)
}
\arguments{
\item{dataset_id}{Character. The dataset identifier.}
}
\description{
Opens the CDC data portal page for a dataset in your default web browser.
}
\examples{
\dontrun{
cdc_browse("swc5-untb")
}

}
