% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cdc_columns}
\alias{cdc_columns}
\title{Get Column Names for a Dataset}
\usage{
cdc_columns(dataset_id, ...)
}
\arguments{
\item{dataset_id}{Character. The dataset identifier.}
}
\value{
Character vector of column names.
}
\description{
Returns the column (field) names for a dataset. Useful for building queries.
}
\examples{
\dontrun{
cdc_columns("swc5-untb")
}

}
